Cast as (gfloat) io->scrollbar_adjustment->value=(gfloat)first_interval;
svn path=/trunk/; revision=22653

+++ gtk/io_stat.c
@@ -950,7 +950,7 @@ io_stat_draw(io_stat_t *io)
 	io->scrollbar_adjustment->step_increment=(gfloat) ((last_interval-first_interval)/10);
 	io->scrollbar_adjustment->page_increment=(gfloat) (last_interval-first_interval);
 	io->scrollbar_adjustment->page_size=io->scrollbar_adjustment->page_increment;
-	io->scrollbar_adjustment->value=first_interval;
+	io->scrollbar_adjustment->value=(gfloat)first_interval;
 	gtk_adjustment_changed(io->scrollbar_adjustment);
 	gtk_adjustment_value_changed(io->scrollbar_adjustment);
